requirements.txt
Django==2.2.1
djangorestframework==3.9.3
django-grappelli==2.12.2
psycopg2-binary==2.8.2
pytest==6.2.5
pytest-django==4.4.0
factory-boy==3.2.0
pytest-mock==3.6.1
